pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Bruno Cestari",email = "brunobcestari@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.5,<23.0)",
    "routeros-api (>=0.21.0,<0.22.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.0.0)",
    "pytest-asyncio (>=0.24.0)",
    "pytest-cov (>=6.0.0)",
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
